requirements.txt
--index-url  https://pypi.douban.com/simple


tensorflow
pillow
pathlib